requirements.txt
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord-py-slash-command==3.0.1
discord.py==1.7.3
future==0.18.2
idna==3.2
importlib-metadata==4.6.4
multidict==5.1.0
pefile==2021.5.24
pyinstaller-hooks-contrib==2021.2
python-dotenv==0.19.0
pywin32-ctypes==0.2.0
typing-extensions==3.10.0.0
yarl==1.6.3
zipp==3.5.0
